requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
redis==5.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
pylint==3.0.2
